before sound(?)

+++ src/Game/MTRANS.c
@@ -1608,14 +1608,15 @@ void seqsAfterProcess() {
                 flSetRenderState(FLRENDER_TEXSTAGE0, val);
 
                 vertices = (TextureVertex*)sceGuGetMemory(2 * sizeof(TextureVertex));
+                //static TextureVertex vertices[2];
                 tex = &flTexture[LO_16_BITS(val) - 1];
 
                 for (int j = 0; j < 2; j++) {
                     vertices[j].x = seqs_w.chip[i].v[j].x;
                     vertices[j].y = seqs_w.chip[i].v[j].y;
                     vertices[j].z = seqs_w.chip[i].v[j].z * 0xFFFF;
-                    vertices[j].u = seqs_w.chip[i].t[j].s * tex->width;
-                    vertices[j].v = seqs_w.chip[i].t[j].t * tex->height;
+                    vertices[j].u = (short) (seqs_w.chip[i].t[j].s * tex->width);
+                    vertices[j].v = (short) (seqs_w.chip[i].t[j].t * tex->height);
                     vertices[j].colour = seqs_w.chip[i].vertex_color;
                 }
                 //ps2SeqsRenderQuad_Ax(&seqs_w.chip[i]);
